updated several protocols and dependencies graph is built, also used to check when deleted
git-svn-id: file:///home/roberto/Git/svn@12209 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 60dcd6c78446ff2c820c7f04043e27003b703bb3

+++ trunk/xmipp/protocols/batch_protocols.py
@@ -597,12 +597,17 @@ class XmippProjectGUI():
                 if state in [SqliteDb.RUN_STARTED, SqliteDb.RUN_LAUNCHED]:
                     showWarning("Delete warning", "This RUN is LAUNCHED or RUNNING, you need to stopped it before delete", parent=self.root) 
                 else:
-                    if askYesNo("Confirm DELETE", "<ALL DATA> related to this <protocol run> will be <DELETED>. \nDo you really want to continue?", self.root):
-                        error = self.project.deleteRun(run)
-                        if error is None:
-                            self.updateRunHistory(self.lastDisplayGroup)
-                        else: 
-                            showError("Error deleting RUN", error, parent=self.root)
+                    runsDict = self.project.getRunsDependencies()
+                    deps = ['<%s>' % d for d in runsDict[getExtendedRunName(run)].deps]
+                    if len(deps):
+                        showWarning("Delete warning", "This RUN is referenced from: \n" + '  \n'.join(deps), parent=self.root) 
+                    else:
+                        if askYesNo("Confirm DELETE", "<ALL DATA> related to this <protocol run> will be <DELETED>. \nDo you really want to continue?", self.root):
+                            error = self.project.deleteRun(run)
+                            if error is None:
+                                self.updateRunHistory(self.lastDisplayGroup)
+                            else: 
+                                showError("Error deleting RUN", error, parent=self.root)
             elif event == ACTION_STEPS:
                 self.launchRunStepsTree(run)
             elif event == ACTION_REFRESH:
